pyproject.toml
[project]
name = "vonage-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[turn-detector]>=1.2.15",
    "livekit-plugins-cartesia>=1.2.15",
    "livekit-plugins-deepgram>=1.2.15",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.2.15",
    "livekit-plugins-silero>=1.2.15",
    "livekit-plugins-turn-detector>=1.2.15",
    "livekit-plugins-google>=1.2.15",
    "livekit-plugins-elevenlabs>=1.2.15",
    "livekit-plugins-rime>=1.2.15",
    "livekit-plugins-smallestai>=1.2.15",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "google-cloud-storage>=2.10.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.8",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "itsdangerous>=2.2.0",
    "cryptography>=43.0.0",
    "tzdata>=2024.1",
    "google-genai>=1.46.0",
    "asyncpg>=0.31.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "livekit-plugins-groq>=1.3.6",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "msal>=1.25.0",
    "httpx>=0.27.0",
    "livekit-plugins-sarvam>=1.3",
    "livekit-plugins-ultravox>=1.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
